@@ -1,25 +1,23 @@
 package com.talhanation.smallships.client.render;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
-import com.mojang.blaze3d.vertex.IVertexBuilder;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import com.mojang.math.Vector3f;
 import com.talhanation.smallships.Main;
 import com.talhanation.smallships.client.model.ModelCogSail;
 import com.talhanation.smallships.client.model.ModelSail;
 import com.talhanation.smallships.entities.AbstractSailShip;
 import com.talhanation.smallships.entities.CogEntity;
-import net.minecraft.client.renderer.IRenderTypeBuffer;
-import net.minecraft.client.renderer.model.ModelRenderer;
+import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.texture.OverlayTexture;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.vector.Vector3f;
+import net.minecraft.resources.ResourceLocation;
 
 public class RenderSailColor {
-
-    public static void renderSailColor(AbstractSailShip ship, float partialTicks, MatrixStack matrixStackIn, String sailColor ,IRenderTypeBuffer bufferIn,  int packedLight, ModelSail modelSail) {
+    public static void renderSailColor(AbstractSailShip ship, float partialTicks, PoseStack matrixStackIn, String sailColor ,MultiBufferSource bufferIn,  int packedLight, ModelSail modelSail) {
         matrixStackIn.pushPose();
-        modelSail.setupAnim(ship, partialTicks, 0.0F, -0.1F, 0.0F, 0.0F);
-        IVertexBuilder ivertexbuilder = bufferIn.getBuffer(modelSail.renderType(getSailColor(sailColor)));
-        modelSail.renderToBuffer(matrixStackIn, ivertexbuilder, packedLight, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
+        model.setupAnim(ship, partialTicks, 0.0F, -0.1F, 0.0F, 0.0F);
+        VertexConsumer vertexConsumer = bufferIn.getBuffer(modelSail.renderType(getSailColor(sailColor)));
+        modelSail.renderToBuffer(matrixStackIn, vertexConsumer, packedLight, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
         matrixStackIn.mulPose(Vector3f.YP.rotationDegrees(-90F));
         matrixStackIn.popPose();
     }
@@ -43,57 +41,25 @@ public class RenderSailColor {
 
 
     private static ResourceLocation getSailColor(String color) {
-        switch (color) {
-            default:
-            case "white":
-                return WHITE;
-
-            case "orange":
-                return ORANGE;
-
-            case "magenta":
-                return MAGENTA;
-
-            case "light_blue":
-                return LIGHT_BLUE;
-
-            case "yellow":
-                return YELLOW;
-
-            case "lime":
-                return LIME;
-
-            case "pink":
-                return PINK;
-
-            case "purple":
-                return PURPLE;
-
-            case "gray":
-                return GRAY;
-
-            case "light_gray":
-                return LIGHT_GRAY;
-
-            case "cyan":
-                return CYAN;
-
-            case "blue":
-                return BLUE;
-
-            case "brown":
-                return BROWN;
-
-            case "green":
-                return GREEN;
-
-            case "red":
-                return RED;
-
-            case "black":
-                return BLACK;
-
-        }
+        return switch (color) {
+            case "white" -> WHITE;
+            case "orange" -> ORANGE;
+            case "magenta" -> MAGENTA;
+            case "light_blue" -> LIGHT_BLUE;
+            case "yellow" -> YELLOW;
+            case "lime" -> LIME;
+            case "pink" -> PINK;
+            case "purple" -> PURPLE;
+            case "gray" -> GRAY;
+            case "light_gray" -> LIGHT_GRAY;
+            case "cyan" -> CYAN;
+            case "blue" -> BLUE;
+            case "brown" -> BROWN;
+            case "green" -> GREEN;
+            case "red" -> RED;
+            case "black" -> BLACK;
+            default -> WHITE;
+        };
     }
 
 }
