@@ -1,6 +1,6 @@
 package com.talhanation.smallships.entities;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.PoseStack;
 import com.talhanation.smallships.DamageSourceShip;
 import com.talhanation.smallships.Main;
 import com.talhanation.smallships.client.model.ModelSail;
@@ -10,27 +10,28 @@ import com.talhanation.smallships.init.SoundInit;
 import com.talhanation.smallships.network.MessageControlShip;
 import com.talhanation.smallships.network.MessageSailState;
 import de.maxhenkel.corelib.math.MathUtils;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.LilyPadBlock;
-import net.minecraft.client.renderer.IRenderTypeBuffer;
-import net.minecraft.entity.*;
-import net.minecraft.entity.passive.WaterMobEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.DyeColor;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.network.datasync.DataParameter;
-import net.minecraft.network.datasync.DataSerializers;
-import net.minecraft.network.datasync.EntityDataManager;
-import net.minecraft.util.*;
-import net.minecraft.util.math.AxisAlignedBB;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.world.World;
-import net.minecraft.world.biome.Biome;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.EntityDataSerializers;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.util.Mth;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.*;
+import net.minecraft.world.entity.animal.WaterAnimal;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.DyeColor;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.WaterlilyBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.AABB;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
@@ -44,20 +45,20 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
     private float prevWaveAngle;
     private boolean collidedLastTick;
 
-    private static final DataParameter<Float> SPEED = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.FLOAT);
-    private static final DataParameter<Float> ROT_SPEED = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.FLOAT);
-    private static final DataParameter<Boolean> FORWARD = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.BOOLEAN);
-    private static final DataParameter<Boolean> BACKWARD = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.BOOLEAN);
-    private static final DataParameter<Boolean> LEFT = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.BOOLEAN);
-    private static final DataParameter<Boolean> RIGHT = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.BOOLEAN);
-    private static final DataParameter<Integer> SAIL_STATE = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.INT);
-    private static final DataParameter<String>  SAIL_COLOR = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.STRING);
+    private static final EntityDataAccessor<Float> SPEED = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.FLOAT);
+    private static final EntityDataAccessor<Float> ROT_SPEED = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.FLOAT);
+    private static final EntityDataAccessor<Boolean> FORWARD = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.BOOLEAN);
+    private static final EntityDataAccessor<Boolean> BACKWARD = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.BOOLEAN);
+    private static final EntityDataAccessor<Boolean> LEFT = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.BOOLEAN);
+    private static final EntityDataAccessor<Boolean> RIGHT = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.BOOLEAN);
+    private static final EntityDataAccessor<Integer> SAIL_STATE = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.INT);
+    private static final EntityDataAccessor<String>  SAIL_COLOR = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.STRING);
 
-    private static final DataParameter<Integer> TYPE = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.INT);
-    private static final DataParameter<Boolean> LEFT_PADDLE = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.BOOLEAN);
-    private static final DataParameter<Boolean> RIGHT_PADDLE = EntityDataManager.defineId(AbstractSailShip.class, DataSerializers.BOOLEAN);
+    private static final EntityDataAccessor<Integer> TYPE = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.INT);
+    private static final EntityDataAccessor<Boolean> LEFT_PADDLE = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.BOOLEAN);
+    private static final EntityDataAccessor<Boolean> RIGHT_PADDLE = SynchedEntityData.defineId(AbstractSailShip.class, EntityDataSerializers.BOOLEAN);
 
-    public AbstractSailShip(EntityType<? extends AbstractWaterVehicle> type, World world) {
+    public AbstractSailShip(EntityType<? extends AbstractWaterVehicle> type, Level world) {
         super(type, world);
         this.maxUpStep = 0.2F;
     }
@@ -83,9 +84,9 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
     public abstract float getVelocityResistance();
     public abstract boolean getHasBanner();
     public abstract void  WaterSplash();
-    public abstract void onInteractionWithShears(PlayerEntity player);
+    public abstract void onInteractionWithShears(Player player);
     public abstract void onCannonKeyPressed();
-    public abstract boolean onInteractionWithBanner(ItemStack itemStack,PlayerEntity player);
+    public abstract boolean onInteractionWithBanner(ItemStack itemStack,Player player);
 
     ////////////////////////////////////TICK////////////////////////////////////
 
@@ -97,8 +98,8 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
 
 
         if ((getSpeed() > 0.085F || getSpeed() < -0.085F)) {
-            this.knockBack(this.level.getEntities(this, this.getBoundingBox().inflate(4.0D, 2.0D, 4.0D).move(0.0D, -2.0D, 0.0D), EntityPredicates.NO_CREATIVE_OR_SPECTATOR));
-            this.knockBack(this.level.getEntities(this, this.getBoundingBox().inflate(4.0D, 2.0D, 4.0D).move(0.0D, -2.0D, 0.0D), EntityPredicates.NO_CREATIVE_OR_SPECTATOR));
+            this.knockBack(this.level.getEntities(this, this.getBoundingBox().inflate(4.0D, 2.0D, 4.0D).move(0.0D, -2.0D, 0.0D), EntitySelector.NO_CREATIVE_OR_SPECTATOR));
+            this.knockBack(this.level.getEntities(this, this.getBoundingBox().inflate(4.0D, 2.0D, 4.0D).move(0.0D, -2.0D, 0.0D), EntitySelector.NO_CREATIVE_OR_SPECTATOR));
 
             if (this.getStatus() == Status.IN_WATER) {
 
@@ -111,7 +112,7 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
 
         }
 
-        if ((this.getControllingPassenger() == null ||!(this.getControllingPassenger() instanceof PlayerEntity) )) {
+        if ((this.getControllingPassenger() == null ||!(this.getControllingPassenger() instanceof Player) )) {
             setSailState(0);
         }
 
@@ -128,14 +129,14 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
     ////////////////////////////////////SAVE DATA////////////////////////////////////
 
     @Override
-    public void addAdditionalSaveData(CompoundNBT nbt) {
+    public void addAdditionalSaveData(CompoundTag nbt) {
 
         nbt.putString("SailColor", getSailColor());
         nbt.putString("Type", getWoodType().getName());
     }
 
     @Override
-    public void readAdditionalSaveData(CompoundNBT nbt) {;
+    public void readAdditionalSaveData(CompoundTag nbt) {
         this.setSailColor(nbt.getString("SailColor"));
         if (nbt.contains("Type", 8)) {
             this.setWoodType(AbstractSailShip.Type.getTypeFromString(nbt.getString("Type")));
@@ -198,7 +199,7 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
     }
 
     public float getWaveAngle(float partialTicks) {
-        return MathHelper.lerp(partialTicks, this.prevWaveAngle, this.waveAngle);
+        return Mth.lerp(partialTicks, this.prevWaveAngle, this.waveAngle);
     }
 
     public boolean isForward() {
@@ -224,13 +225,12 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
     }
 
     public boolean isAccelerating() {
-        boolean b = (isForward() || isBackward()) && !horizontalCollision;
-        return b;
+        return (isForward() || isBackward()) && !horizontalCollision;
     }
 
     @OnlyIn(Dist.CLIENT)
     public float getRowingTime(int side, float limbSwing) {
-        return this.getPaddleState(side) ? (float) MathHelper.clampedLerp((double) this.paddlePositions[side] - (double) ((float) Math.PI / 8F), (double) this.paddlePositions[side], (double) limbSwing) : 0.0F;
+        return this.getPaddleState(side) ? (float) Mth.clampedLerp((double) this.paddlePositions[side] - (double) ((float) Math.PI / 8F), this.paddlePositions[side], limbSwing) : 0.0F;
     }
 
     ////////////////////////////////////SET////////////////////////////////////
@@ -287,9 +287,9 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
         }
     }
 
-    public void sendSteerStateToServer(){
-        //Main.SIMPLE_CHANNEL.sendToServer(new MessageSteerState(this.getSteerState(0), this.getSteerState(1)));
-    }
+    /*public void sendSteerStateToServer(){
+        Main.SIMPLE_CHANNEL.sendToServer(new MessageSteerState(this.getSteerState(0), this.getSteerState(1)));
+    }*/
 
     @Override
     public boolean canCollideWith(Entity entityIn) {
@@ -309,13 +309,13 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
     }
 
     public void checkPush() {
-        List<PlayerEntity> list = level.getEntitiesOfClass(PlayerEntity.class, getBoundingBox().expandTowards(0.2, 0, 0.2).expandTowards(-0.2, 0, -0.2));
+        List<Player> list = level.getEntitiesOfClass(Player.class, getBoundingBox().expandTowards(0.2, 0, 0.2).expandTowards(-0.2, 0, -0.2));
 
-        for (PlayerEntity player : list) {
+        for (Player player : list) {
             if (!player.hasPassenger(this) && player.isShiftKeyDown()) {
-                double motX = calculateMotionX(0.05F, player.yRot);
-                double motZ = calculateMotionZ(0.05F, player.yRot);
-                move(MoverType.PLAYER, new Vector3d(motX, 0, motZ));
+                double motX = calculateMotionX(0.05F, player.getYRot());
+                double motZ = calculateMotionZ(0.05F, player.getYRot());
+                move(MoverType.PLAYER, new Vec3(motX, 0, motZ));
                 return;
             }
         }
@@ -344,28 +344,27 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
 
         if (sailstate != 0) {
             switch (sailstate) {
-                case 1:
-                    maxSp *= 4/16F;
+                case 1 -> {
+                    maxSp *= 4 / 16F;
                     if (speed <= maxSp)
                         speed = Math.min(speed + getAcceleration() * 2F / 8, maxSp);
-                    break;
-                case 2:
-                    maxSp *= 8/16F;
+                }
+                case 2 -> {
+                    maxSp *= 8 / 16F;
                     if (speed <= maxSp)
                         speed = Math.min(speed + getAcceleration() * 3.5F / 8, maxSp);
-                    break;
-
-                case 3:
-                    maxSp *= 12/16F;
+                }
+                case 3 -> {
+                    maxSp *= 12 / 16F;
                     if (speed <= maxSp)
                         speed = Math.min(speed + getAcceleration() * 5 / 8, maxSp);
-                    break;
-                case 4:
+                }
+                case 4 -> {
                     maxSp *= 1F;
                     if (speed <= maxSp) {
                         speed = Math.min(speed + getAcceleration(), maxSp);
                     }
-                    break;
+                }
             }
         }
 
@@ -403,9 +402,9 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
 
         deltaRotation = rotationSpeed;
 
-        yRot += deltaRotation;
+        setYRot(getYRot() + deltaRotation);
 
-        setDeltaMovement(calculateMotionX(getSpeed(), yRot), getDeltaMovement().y, calculateMotionZ(getSpeed(), yRot));
+        setDeltaMovement(calculateMotionX(getSpeed(), getYRot()), getDeltaMovement().y, calculateMotionZ(getSpeed(), getYRot()));
     }
 
     ////////////////////////////////////ON FUNCTIONS////////////////////////////////////
@@ -445,7 +444,7 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
     }
 
 
-    public void onInteractionWithDye(PlayerEntity player, DyeColor dyeColor, ItemStack itemStack) {
+    public void onInteractionWithDye(Player player, DyeColor dyeColor, ItemStack itemStack) {
         String color = dyeColor.getName();
         setSailColor(color);
         if (!player.isCreative()) {
@@ -471,8 +470,8 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
     public void updateWaterMobs() {
         if (SmallShipsConfig.WaterMobFlee.get()) {
             double radius = 15.0D;
-            List<WaterMobEntity> list1 = this.level.getEntitiesOfClass(WaterMobEntity.class, new AxisAlignedBB(getX() - radius, getY() - radius, getZ() - radius, getX() + radius, getY() + radius, getZ() + radius));
-            for (WaterMobEntity ent : list1)
+            List<WaterAnimal> list1 = this.level.getEntitiesOfClass(WaterAnimal.class, new AABB(getX() - radius, getY() - radius, getZ() - radius, getX() + radius, getY() + radius, getZ() + radius));
+            for (WaterAnimal ent : list1)
                 fleeEntity(ent);
         }
     }
@@ -496,18 +495,18 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
                 momentum = 0.9F;
             }
 
-            Vector3d vector3d = this.getDeltaMovement();
+            Vec3 vector3d = this.getDeltaMovement();
             this.setDeltaMovement(vector3d.x * (double) momentum, vector3d.y + d1, vector3d.z * (double) momentum);
             this.deltaRotation *= momentum;
             if (d2 > 0.0D) {
-                Vector3d vector3d1 = this.getDeltaMovement();
+                Vec3 vector3d1 = this.getDeltaMovement();
                 this.setDeltaMovement(vector3d1.x, (vector3d1.y + d2 * 0.06153846016296973D) * 0.75D, vector3d1.z);
             }
         }
 
     }
 
-    public void updateControls(boolean forward, boolean backward, boolean left, boolean right, PlayerEntity player) {
+    public void updateControls(boolean forward, boolean backward, boolean left, boolean right, Player player) {
         boolean needsUpdate = false;
 
         if (isForward() != forward) {
@@ -536,7 +535,7 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
 
     ////////////////////////////////////OTHER FUNCTIONS////////////////////////////////////
 
-    public boolean canPlayerEnterShip(PlayerEntity player) {
+    public boolean canPlayerEnterShip(Player player) {
         return true;
     }
 
@@ -552,21 +551,21 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
 
     ////////////////////////////////////OTHER FUNCTIONS////////////////////////////////////
 
-    public void renderSailColor(MatrixStack matrixStack, IRenderTypeBuffer buffer , int packedLight, float partialTicks, ModelSail sailModel) {
+    public void renderSailColor(PoseStack matrixStack, MultiBufferSource buffer , int packedLight, float partialTicks, ModelSail sailModel) {
         RenderSailColor.renderSailColor(this, partialTicks, matrixStack, getSailColor(), buffer,  packedLight, sailModel);
     }
 
-    public void destroyShip(DamageSource dmg) {
-        remove();
+    public void destroyShip(DamageSource src) {
+        remove(RemovalReason.KILLED);
     }
 
-    public void fleeEntity(MobEntity entity) {
+    public void fleeEntity(Mob entity) {
         double fleeDistance = 10.0D;
-        Vector3d vecBoat = new Vector3d(getX(), getY(), getZ());
-        Vector3d vecEntity = new Vector3d(entity.getX(), entity.getY(), entity.getZ());
-        Vector3d fleeDir = vecEntity.subtract(vecBoat);
+        Vec3 vecBoat = new Vec3(getX(), getY(), getZ());
+        Vec3 vecEntity = new Vec3(entity.getX(), entity.getY(), entity.getZ());
+        Vec3 fleeDir = vecEntity.subtract(vecBoat);
         fleeDir = fleeDir.normalize();
-        Vector3d fleePos = new Vector3d(vecEntity.x + fleeDir.x * fleeDistance, vecEntity.y + fleeDir.y * fleeDistance, vecEntity.z + fleeDir.z * fleeDistance);
+        Vec3 fleePos = new Vec3(vecEntity.x + fleeDir.x * fleeDistance, vecEntity.y + fleeDir.y * fleeDistance, vecEntity.z + fleeDir.z * fleeDistance);
         entity.getNavigation().moveTo(fleePos.x, fleePos.y, fleePos.z, 1.5D);
     }
 
@@ -579,17 +578,17 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
                 double d2 = entity.getX() - d0;
                 double d3 = entity.getZ() - d1;
                 double d4 = Math.max(d2 * d2 + d3 * d3, 0.1D);
-                entity.push(d2 / d4 * 0.4D, (double)0.0F, d3 / d4 * 0.4D);
+                entity.push(d2 / d4 * 0.4D, 0.0F, d3 / d4 * 0.4D);
             }
         }
     }
 
     private void breakLily() {
-        AxisAlignedBB boundingBox = getBoundingBox();
+        AABB boundingBox = getBoundingBox();
         double offset = 0.75D;
         BlockPos start = new BlockPos(boundingBox.minX - offset, boundingBox.minY - offset, boundingBox.minZ - offset);
         BlockPos end = new BlockPos(boundingBox.maxX + offset, boundingBox.maxY + offset, boundingBox.maxZ + offset);
-        BlockPos.Mutable pos = new BlockPos.Mutable();
+        BlockPos.MutableBlockPos pos = new BlockPos.MutableBlockPos();
         boolean hasBroken = false;
         if (level.hasChunksAt(start, end)) {
             for (int i = start.getX(); i <= end.getX(); ++i) {
@@ -597,7 +596,7 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
                     for (int k = start.getZ(); k <= end.getZ(); ++k) {
                         pos.set(i, j, k);
                         BlockState blockstate = level.getBlockState(pos);
-                        if (blockstate.getBlock() instanceof LilyPadBlock) {
+                        if (blockstate.getBlock() instanceof WaterlilyBlock) {
                             level.destroyBlock(pos, true);
                             hasBroken = true;
                         }
@@ -606,26 +605,19 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
             }
         }
         if (hasBroken) {
-            level.playSound(null, getX(), getY(), getZ(), SoundEvents.CROP_BREAK, SoundCategory.BLOCKS, 1F, 0.9F + 0.2F * random.nextFloat());
+            level.playSound(null, getX(), getY(), getZ(), SoundEvents.CROP_BREAK, SoundSource.BLOCKS, 1F, 0.9F + 0.2F * random.nextFloat());
         }
     }
 
     public Item getItemBoat() {
-        switch (this.getWoodType()) {
-            case OAK:
-            default:
-                return Items.OAK_BOAT;
-            case SPRUCE:
-                return Items.SPRUCE_BOAT;
-            case BIRCH:
-                return Items.BIRCH_BOAT;
-            case JUNGLE:
-                return Items.JUNGLE_BOAT;
-            case ACACIA:
-                return Items.ACACIA_BOAT;
-            case DARK_OAK:
-                return Items.DARK_OAK_BOAT;
-        }
+        return switch (this.getWoodType()) {
+            case OAK -> Items.OAK_BOAT;
+            case SPRUCE -> Items.SPRUCE_BOAT;
+            case BIRCH -> Items.BIRCH_BOAT;
+            case JUNGLE -> Items.JUNGLE_BOAT;
+            case ACACIA -> Items.ACACIA_BOAT;
+            case DARK_OAK -> Items.DARK_OAK_BOAT;
+        };
     }
 
     public enum Type {
@@ -705,9 +697,9 @@ public abstract class AbstractSailShip extends AbstractWaterVehicle {
         public static AbstractSailShip.Type getTypeFromString(String nameIn) {
             AbstractSailShip.Type[] aboatentity$type = values();
 
-            for (int i = 0; i < aboatentity$type.length; ++i) {
-                if (aboatentity$type[i].getName().equals(nameIn)) {
-                    return aboatentity$type[i];
+            for (Type type : aboatentity$type) {
+                if (type.getName().equals(nameIn)) {
+                    return type;
                 }
             }
 
