@@ -3,33 +3,47 @@ package com.talhanation.smallships.entities;
 import com.talhanation.smallships.InventoryEvents;
 import com.talhanation.smallships.init.ModEntityTypes;
 import com.talhanation.smallships.init.ModItems;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.*;
-import net.minecraft.particles.ParticleTypes;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.world.World;
+import com.talhanation.smallships.inventory.BasicShipContainer;
+import com.talhanation.smallships.network.MessageOpenGui;
+import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.EntityDataSerializers;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.util.Mth;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.item.*;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.fmllegacy.network.NetworkHooks;
+
+import javax.annotation.Nullable;
 
 public class CogEntity extends AbstractCannonShip{
 
-    public CogEntity(EntityType<? extends CogEntity> type, World world) {
+    public CogEntity(EntityType<? extends CogEntity> type, Level world) {
         super(type, world);
     }
 
     //Constructor for ShipItem
-    public CogEntity(World world, double x, double y, double z) {
+    public CogEntity(Level world, double x, double y, double z) {
         this(ModEntityTypes.COG.get(), world);
         setPos(x, y, z);
         this.xo = x;
         this.yo = y;
         this.zo = z;
     }
-
+    
     ////////////////////////////////////GET////////////////////////////////////
 
     // hight and width for now as mast
@@ -43,6 +57,7 @@ public class CogEntity extends AbstractCannonShip{
         return 1.5D;
     }
 
+    @Override
     public double getShipDefense() { //in %
         return 12;
     }
@@ -90,22 +105,14 @@ public class CogEntity extends AbstractCannonShip{
 
     @Override
     public int getPassengerSize() {
-        switch (getTotalCannonCount()){
-            default:
-            case 0:
-            return 5;
-
-            case 1:
-            case 4:
-                return 2;
-
-            case 2:
-                return 3;
-
-            case 3:
-                return 3;
-
-        }
+        return switch (getTotalCannonCount()) {
+            case 0 -> 5;
+            case 1 -> 2;
+            case 2 -> 3;
+            case 3 -> 3;
+            case 4 -> 2;
+            default -> throw new IllegalStateException("Unexpected passenger size: " + getTotalCannonCount());
+        };
     }
 
     public int getMaxCannons(){//max cannons
@@ -115,17 +122,18 @@ public class CogEntity extends AbstractCannonShip{
     ////////////////////////////////////INTERACTIONS///////////////////////////////
 
     @Override
-    public ActionResultType interact(PlayerEntity player, Hand hand) {
+    public InteractionResult interact(Player player, InteractionHand hand) {
+
         ItemStack itemInHand = player.getItemInHand(hand);
         if (player.isSecondaryUseActive()) {
 
-            if (this.isVehicle() && !(getControllingPassenger() instanceof PlayerEntity)) {
+            if (this.isVehicle() && !(getControllingPassenger() instanceof Player)) {
                 this.ejectPassengers();
             } else {
-                if (!(getControllingPassenger() instanceof PlayerEntity)) {
+                if (!(getControllingPassenger() instanceof Player)) {
                     InventoryEvents.openShipGUI(player, this,0);
                 }
-                return ActionResultType.sidedSuccess(this.level.isClientSide);
+                return InteractionResult.sidedSuccess(this.level.isClientSide);
             }
 
         }
@@ -133,43 +141,44 @@ public class CogEntity extends AbstractCannonShip{
         if (!this.getSunken()) {
             if (itemInHand.getItem() == ModItems.CANNON_ITEM.get()) {
                 this.onInteractionWithCannon(player, itemInHand);
-                return ActionResultType.SUCCESS;
+                return InteractionResult.SUCCESS;
             }
 
             if (itemInHand.getItem() instanceof DyeItem) {
                 this.onInteractionWithDye(player, ((DyeItem) itemInHand.getItem()).getDyeColor(), itemInHand);
-                return ActionResultType.SUCCESS;
+                return InteractionResult.SUCCESS;
             }
 
             if (itemInHand.getItem() instanceof BannerItem) {
                 this.onInteractionWithBanner(itemInHand, player);
-                return ActionResultType.SUCCESS;
+                return InteractionResult.SUCCESS;
             }
+            if (!player.isSecondaryUseActive()) {
 
             if (itemInHand.getItem() instanceof AxeItem) {
                 if (hasPlanks(player.inventory) && hasIronNugget(player.inventory) && getShipDamage() > 16.0D) {
                     this.onInteractionWitAxe(player);
-                    return ActionResultType.SUCCESS;
-                } else return ActionResultType.FAIL;
+                    return InteractionResult.SUCCESS;
+                } else return InteractionResult.FAIL;
             } else if (itemInHand.getItem() instanceof ShearsItem) {
                 if (this.getHasBanner()) {
                     this.onInteractionWithShears(player);
-                    return ActionResultType.SUCCESS;
+                    return InteractionResult.SUCCESS;
                 }
-                return ActionResultType.PASS;
+                return InteractionResult.PASS;
             }
             if (!player.isSecondaryUseActive()) {
 
                 if (!this.level.isClientSide) {
-                    return player.startRiding(this) ? ActionResultType.CONSUME : ActionResultType.PASS;
+                    return player.startRiding(this) ? InteractionResult.CONSUME : InteractionResult.PASS;
 
                 } else {
-                    return ActionResultType.SUCCESS;
+                    return InteractionResult.SUCCESS;
                 }
             }
         }
 
-       return ActionResultType.FAIL;
+       return InteractionResult.FAIL;
     }
 
 
@@ -182,11 +191,11 @@ public class CogEntity extends AbstractCannonShip{
     ////////////////////////////////////OTHER FUNCTIONS////////////////////////////////////
     @Override
     public void WaterSplash(){
-        Vector3d vector3d = this.getViewVector(0.0F);
-        float f0 = MathHelper.cos(this.yRot * ((float)Math.PI / 180F)) * 0.8F;
-        float f1 = MathHelper.sin(this.yRot * ((float)Math.PI / 180F)) * 0.8F;
-        float f0_1 = MathHelper.cos(this.yRot * ((float)Math.PI / 180F)) * 1.6F;
-        float f1_1 = MathHelper.sin(this.yRot * ((float)Math.PI / 180F)) * 1.6F;
+        Vec3 vector3d = this.getViewVector(0.0F);
+        float f0 = Mth.cos(this.getYRot() * ((float)Math.PI / 180F)) * 0.8F;
+        float f1 = Mth.sin(this.getYRot() * ((float)Math.PI / 180F)) * 0.8F;
+        float f0_1 = Mth.cos(this.getYRot() * ((float)Math.PI / 180F)) * 1.6F;
+        float f1_1 = Mth.sin(this.getYRot() * ((float)Math.PI / 180F)) * 1.6F;
         float f2 =  2.5F - this.random.nextFloat() * 0.7F;
         float f2_ =  -1.3F - this.random.nextFloat() * 0.7F;
         float x = 0;
@@ -221,17 +230,15 @@ public class CogEntity extends AbstractCannonShip{
             float f = -1.75F; //driver x pos
             float d = 0.0F;   //driver z pos
             float x = 0;//global offset
-            float f1 = (float) ((this.removed ? 0.02D : getPassengersRidingOffset()) + passenger.getMyRidingOffset());
+            float f1 = (float) ((this.isRemoved() ? 0.02D : getPassengersRidingOffset()) + passenger.getMyRidingOffset());
             if (getPassengers().size() == 2) {
                 int i = getPassengers().indexOf(passenger);
                 if (i == 0) {
-
                     f = -1.75F;
-                    d = 0.0F;
                 } else {
                     f = 1.25F;
-                    d = 0.0F;
                 }
+                d = 0.0F;
             } else if (getPassengers().size() == 3) {
                 int i = getPassengers().indexOf(passenger);
                 if (i == 0) {
@@ -279,11 +286,11 @@ public class CogEntity extends AbstractCannonShip{
                 }
             }
             f = f + x;
-            Vector3d vector3d = (new Vector3d((double)f, 0.0D, 0.0D + d)).yRot(-this.yRot * ((float)Math.PI / 180F) - ((float)Math.PI / 2F));
-            passenger.setPos(this.getX() + vector3d.x, this.getY() + (double)f1, + this.getZ() + vector3d.z);
-            passenger.yRot += this.deltaRotation;
+            Vec3 vector3d = (new Vec3(f, 0.0D, 0.0D + d)).yRot(-this.getYRot() * ((float)Math.PI / 180F) - ((float)Math.PI / 2F));
+            passenger.setPos(this.getX() + vector3d.x, this.getY() + (double)f1, this.getZ() + vector3d.z);
+            passenger.setYRot(passenger.getYRot() + this.deltaRotation);
             passenger.setYHeadRot(passenger.getYHeadRot() + this.deltaRotation);
-            applyYawToEntity(passenger);
+            applyOriantationsToEntity(passenger);
         }
 
     }
