@@ -1,6 +1,6 @@
 package com.talhanation.smallships;
 
-import net.minecraft.util.DamageSource;
+import net.minecraft.world.damagesource.DamageSource;
 
 public class DamageSourceShip extends DamageSource {
 
