@@ -1,18 +1,18 @@
 package com.talhanation.smallships.client.gui;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.talhanation.smallships.Main;
 import com.talhanation.smallships.entities.AbstractInventoryEntity;
+import com.mojang.blaze3d.vertex.PoseStack;
 import com.talhanation.smallships.entities.AbstractShipDamage;
 import com.talhanation.smallships.inventory.BasicShipContainer;
 import com.talhanation.smallships.network.MessageOpenGui;
 import de.maxhenkel.corelib.inventory.ScreenBase;
 import net.minecraft.client.gui.widget.button.Button;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.StringTextComponent;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Inventory;
 
 public class BasicShipInvScreen extends ScreenBase<BasicShipContainer> {
 
@@ -20,9 +20,9 @@ public class BasicShipInvScreen extends ScreenBase<BasicShipContainer> {
 
     private final BasicShipContainer container;
     private final AbstractInventoryEntity ship;
-    private final PlayerInventory playerInventory;
+    private final Inventory playerInventory;
 
-    public BasicShipInvScreen(BasicShipContainer container, PlayerInventory playerInventory, ITextComponent title) {
+    public BasicShipInvScreen(BasicShipContainer container, Inventory playerInventory, Component title) {
         super(GUI_TEXTURE, container, playerInventory, title);
         this.ship = container.getShip();
         this.playerInventory = playerInventory;
@@ -39,7 +39,7 @@ public class BasicShipInvScreen extends ScreenBase<BasicShipContainer> {
         int zeroTopPos = topPos + 15;
 
         if (ship.getMaxInvPage() > 1 && ship.getInvPage() > 1){
-            addButton(new Button(zeroLeftPos - 205, zeroTopPos, 40, 20, new StringTextComponent("<-"), button -> {
+            addButton(new Button(zeroLeftPos - 205, zeroTopPos, 40, 20, new TextComponent("<-"), button -> {
 
                 Main.SIMPLE_CHANNEL.sendToServer(new MessageOpenGui(playerInventory.player, ship, 0));
                 ship.setInvPage(ship.getInvPage() - 1);
@@ -47,7 +47,7 @@ public class BasicShipInvScreen extends ScreenBase<BasicShipContainer> {
         }
 
         if(ship.getMaxInvPage() > 1 && ship.getInvPage() < ship.getMaxInvPage()){
-            addButton(new Button(zeroLeftPos + 20, zeroTopPos, 40, 20, new StringTextComponent("->"), button -> {
+            addButton(new Button(zeroLeftPos + 20, zeroTopPos, 40, 20, new TextComponent("->"), button -> {
 
                     ship.setInvPage(ship.getInvPage() + 1);
                     Main.SIMPLE_CHANNEL.sendToServer(new MessageOpenGui(playerInventory.player, ship, 54));
@@ -57,7 +57,7 @@ public class BasicShipInvScreen extends ScreenBase<BasicShipContainer> {
 
 
     @Override
-    protected void renderLabels(MatrixStack matrixStack, int mouseX, int mouseY) {
+    protected void renderLabels(PoseStack matrixStack, int mouseX, int mouseY) {
         super.renderLabels(matrixStack, mouseX, mouseY);
         boolean renderInvPage = ship.getMaxInvPage() != 1;
         font.draw(matrixStack, ship.getDisplayName().getVisualOrderText(), 8, 6, FONT_COLOR);
@@ -71,9 +71,9 @@ public class BasicShipInvScreen extends ScreenBase<BasicShipContainer> {
         return false;
     }
 
-    protected void renderBg(MatrixStack matrixStack, float partialTicks, int mouseX, int mouseY) {
+    protected void renderBg(PoseStack matrixStack, float partialTicks, int mouseX, int mouseY) {
         super.renderBg(matrixStack, partialTicks, mouseX, mouseY);
-        RenderSystem.color4f(1.0F, 1.0F, 1.0F, 1.0F);
+        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
     }
 
 
