@@ -1,41 +1,49 @@
-package com.talhanation.smallships.client.model;// Made with Blockbench 3.9.3
-// Exported for Minecraft version 1.15 - 1.16 with MCP mappings
+package com.talhanation.smallships.client.model;
+
+// Made with Blockbench 4.2.5
+// Exported for Minecraft version 1.17 - 1.18 with Mojang mappings
 // Paste this class into your mod and generate all required imports
 
-
-import com.mojang.blaze3d.matrix.MatrixStack;
-import com.mojang.blaze3d.vertex.IVertexBuilder;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import com.talhanation.smallships.Main;
 import com.talhanation.smallships.entities.projectile.CannonBallEntity;
-import net.minecraft.client.renderer.entity.model.EntityModel;
-import net.minecraft.client.renderer.model.ModelRenderer;
+import net.minecraft.client.model.EntityModel;
+import net.minecraft.client.model.geom.ModelLayerLocation;
+import net.minecraft.client.model.geom.ModelPart;
+import net.minecraft.client.model.geom.PartPose;
+import net.minecraft.client.model.geom.builders.*;
+import net.minecraft.resources.ResourceLocation;
 
 public class ModelCannonBall extends EntityModel<CannonBallEntity> {
-	private final ModelRenderer cannonball;
+	// This layer location should be baked with EntityRendererProvider.Context in the entity renderer and passed into this model's constructor
+	public static final ModelLayerLocation LAYER_LOCATION = new ModelLayerLocation(new ResourceLocation(Main.MOD_ID, "model_cannonball"), "main");
+	private final ModelPart cannonball;
 
 	public ModelCannonBall() {
-		texWidth = 16;
-		texHeight = 16;
-
-		cannonball = new ModelRenderer(this);
-		cannonball.setPos(0.0F, 23.5F, 0.0F);
-		cannonball.setTexSize(0, 0).addBox(1.5F, -4.0F, -1.0F, 1.0F, 2.0F, 2.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-3.0F, -1.0F, -2.0F, 4.0F, 1.0F, 4.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-3.0F, -6.0F, -2.0F, 4.0F, 1.0F, 4.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-2.0F, -6.5F, -1.0F, 2.0F, 1.0F, 2.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-2.0F, -0.5F, -1.0F, 2.0F, 1.0F, 2.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-3.0F, -5.0F, 2.0F, 4.0F, 4.0F, 1.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-2.0F, -4.0F, 2.5F, 2.0F, 2.0F, 1.0F, 0.0F, false);
-		cannonball.setTexSize(10, 0).addBox(-2.0F, -4.0F, -3.5F, 2.0F, 2.0F, 1.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-3.0F, -5.0F, -3.0F, 4.0F, 4.0F, 1.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(1.0F, -5.0F, -2.0F, 1.0F, 4.0F, 4.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-4.0F, -5.0F, -2.0F, 1.0F, 4.0F, 4.0F, 0.0F, false);
-		cannonball.setTexSize(0, 0).addBox(-4.5F, -4.0F, -1.0F, 1.0F, 2.0F, 2.0F, 0.0F, false);
+		ModelPart root = createBodyLayer().bakeRoot();
+		this.cannonball = root.getChild("cannonball");
 	}
 
-	public void setRotationAngle(ModelRenderer modelRenderer, float x, float y, float z) {
-		modelRenderer.xRot = x;
-		modelRenderer.yRot = y;
-		modelRenderer.zRot = z;
+	public static LayerDefinition createBodyLayer() {
+		MeshDefinition meshdefinition = new MeshDefinition();
+		PartDefinition partdefinition = meshdefinition.getRoot();
+
+		PartDefinition cannonball = partdefinition.addOrReplaceChild("cannonball", CubeListBuilder.create()
+				.addBox(1.5F, -4.0F, -1.0F, 1.0F, 2.0F, 2.0F)
+				.addBox(-3.0F, -1.0F, -2.0F, 4.0F, 1.0F, 4.0F)
+				.addBox(-3.0F, -6.0F, -2.0F, 4.0F, 1.0F, 4.0F)
+				.addBox(-2.0F, -6.5F, -1.0F, 2.0F, 1.0F, 2.0F)
+				.addBox(-2.0F, -0.5F, -1.0F, 2.0F, 1.0F, 2.0F)
+				.addBox(-3.0F, -5.0F, 2.0F, 4.0F, 4.0F, 1.0F)
+				.addBox(-2.0F, -4.0F, 2.5F, 2.0F, 2.0F, 1.0F)
+				.addBox(-2.0F, -4.0F, -3.5F, 2.0F, 2.0F, 1.0F)
+				.addBox(-3.0F, -5.0F, -3.0F, 4.0F, 4.0F, 1.0F)
+				.addBox(1.0F, -5.0F, -2.0F, 1.0F, 4.0F, 4.0F)
+				.addBox(-4.0F, -5.0F, -2.0F, 1.0F, 4.0F, 4.0F)
+				.addBox(-4.5F, -4.0F, -1.0F, 1.0F, 2.0F, 2.0F), PartPose.offset(0.0F, 23.5F, 0.0F));
+
+		return LayerDefinition.create(meshdefinition, 16, 16);
 	}
 
 	@Override
@@ -44,7 +52,7 @@ public class ModelCannonBall extends EntityModel<CannonBallEntity> {
 	}
 
 	@Override
-	public void renderToBuffer(MatrixStack matrixStack, IVertexBuilder buffer, int packedLight, int packedOverlay, float red, float green, float blue, float alpha){
-		cannonball.render(matrixStack, buffer, packedLight, packedOverlay);
+	public void renderToBuffer(PoseStack poseStack, VertexConsumer vertexConsumer, int packedLight, int packedOverlay, float red, float green, float blue, float alpha) {
+		cannonball.render(poseStack, vertexConsumer, packedLight, packedOverlay, red, green, blue, alpha);
 	}
 }
