@@ -2,27 +2,44 @@ package com.talhanation.smallships.entities;
 
 import com.talhanation.smallships.InventoryEvents;
 import com.talhanation.smallships.init.ModEntityTypes;
-import com.talhanation.smallships.init.ModItems;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.*;
-import net.minecraft.particles.ParticleTypes;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.world.World;
+import com.talhanation.smallships.inventory.BasicShipContainer;
+import com.talhanation.smallships.network.MessageOpenGui;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.SimpleContainer;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.item.BannerItem;
+import net.minecraft.world.item.DyeItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ShearsItem;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.EntityDataSerializers;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.Mth;
+import net.minecraft.world.phys.Vec3;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.level.Level;
+import net.minecraftforge.fmllegacy.network.NetworkHooks;
+
+import javax.annotation.Nullable;
 
 public class BriggEntity extends AbstractCannonShip{
 
-    public BriggEntity(EntityType<? extends BriggEntity> type, World world) {
+    public BriggEntity(EntityType<? extends BriggEntity> type, Level world) {
         super(type, world);
     }
 
     //Constructor for ShipItem
-    public BriggEntity(World world, double x, double y, double z) {
+    public BriggEntity(Level world, double x, double y, double z) {
         this(ModEntityTypes.BRIGG.get(), world);
         setPos(x, y, z);
         this.xo = x;
@@ -100,7 +117,7 @@ public class BriggEntity extends AbstractCannonShip{
     ////////////////////////////////////INTERACTIONS///////////////////////////////
 
     @Override
-    public ActionResultType interact(PlayerEntity player, Hand hand) {
+    public InteractionResult interact(Player player, InteractionHand hand) {
         ItemStack itemInHand = player.getItemInHand(hand);
         if (player.isSecondaryUseActive()) {
 
@@ -110,7 +127,7 @@ public class BriggEntity extends AbstractCannonShip{
                 if (!(getControllingPassenger() instanceof PlayerEntity)) {
                     InventoryEvents.openShipGUI(player, this,0);
                 }
-                return ActionResultType.sidedSuccess(this.level.isClientSide);
+                return InteractionResult.sidedSuccess(this.level.isClientSide);
             }
 
         }
@@ -118,43 +135,43 @@ public class BriggEntity extends AbstractCannonShip{
         if (!this.getSunken()) {
             if (itemInHand.getItem() == ModItems.CANNON_ITEM.get()) {
                 this.onInteractionWithCannon(player, itemInHand);
-                return ActionResultType.SUCCESS;
+                return InteractionResult.SUCCESS;
             }
 
             if (itemInHand.getItem() instanceof DyeItem) {
                 this.onInteractionWithDye(player, ((DyeItem) itemInHand.getItem()).getDyeColor(), itemInHand);
-                return ActionResultType.SUCCESS;
+                return InteractionResult.SUCCESS;
             }
 
             if (itemInHand.getItem() instanceof BannerItem) {
                 this.onInteractionWithBanner(itemInHand, player);
-                return ActionResultType.SUCCESS;
+                return InteractionResult.SUCCESS;
             }
 
             if (itemInHand.getItem() instanceof AxeItem) {
                 if (hasPlanks(player.inventory) && hasIronNugget(player.inventory) && getShipDamage() > 16.0D) {
                     this.onInteractionWitAxe(player);
-                    return ActionResultType.SUCCESS;
-                } else return ActionResultType.FAIL;
+                    return InteractionResult.SUCCESS;
+                } else return InteractionResult.FAIL;
             } else if (itemInHand.getItem() instanceof ShearsItem) {
                 if (this.getHasBanner()) {
                     this.onInteractionWithShears(player);
-                    return ActionResultType.SUCCESS;
+                    return InteractionResult.SUCCESS;
                 }
-                return ActionResultType.PASS;
+                return InteractionResult.PASS;
             }
             if (!player.isSecondaryUseActive()) {
 
                 if (!this.level.isClientSide) {
-                    return player.startRiding(this) ? ActionResultType.CONSUME : ActionResultType.PASS;
+                    return player.startRiding(this) ? InteractionResult.CONSUME : InteractionResult.PASS;
 
                 } else {
-                    return ActionResultType.SUCCESS;
+                    return InteractionResult.SUCCESS;
                 }
             }
         }
 
-        return ActionResultType.FAIL;
+        return InteractionResult.FAIL;
     }
 
 
@@ -167,9 +184,9 @@ public class BriggEntity extends AbstractCannonShip{
     ////////////////////////////////////OTHER FUNCTIONS////////////////////////////////////
     @Override
     public void WaterSplash(){
-        Vector3d vector3d = this.getViewVector(0.0F);
-        float f0 = MathHelper.cos(this.yRot * ((float)Math.PI / 180F)) * 1.2F;
-        float f1 = MathHelper.sin(this.yRot * ((float)Math.PI / 180F)) * 1.2F;
+        Vec3 vector3d = this.getViewVector(0.0F);
+        float f0 = Mth.cos(this.getYRot() * ((float)Math.PI / 180F)) * 1.2F;
+        float f1 = Mth.sin(this.getYRot() * ((float)Math.PI / 180F)) * 1.2F;
         float f2 =  4F - this.random.nextFloat() * 0.7F; // höhe
         float f2_ =  -2.3F - this.random.nextFloat() * 0.7F;
         float x = 0; //verschiebung nach rechts/links
@@ -202,7 +219,7 @@ public class BriggEntity extends AbstractCannonShip{
         if (hasPassenger(passenger)) {
             double f = -1.5F;
             double d = 0.75F;
-            float f1 = (float) ((this.removed ? 0.02D : getPassengersRidingOffset()) + passenger.getMyRidingOffset());
+            float f1 = (float) ((this.isRemoved() ? 0.02D : getPassengersRidingOffset()) + passenger.getMyRidingOffset());
             if (getPassengers().size() == 2) {
                 int i = getPassengers().indexOf(passenger);
                 if (i == 0) {
@@ -394,11 +411,11 @@ public class BriggEntity extends AbstractCannonShip{
                 }
             }
             f = f - 0.5;
-            Vector3d vector3d = (new Vector3d((double)f, 0.0D, 0.0D + d)).yRot(-this.yRot * ((float)Math.PI / 180F) - ((float)Math.PI / 2F));
-            passenger.setPos(this.getX() + vector3d.x, this.getY() + (double)f1, + this.getZ() + vector3d.z);
-            passenger.yRot += this.deltaRotation;
+            Vec3 vector3d = (new Vec3(f, 0.0D, 0.0D + d)).yRot(-this.getYRot() * ((float)Math.PI / 180F) - ((float)Math.PI / 2F));
+            passenger.setPos(this.getX() + vector3d.x, this.getY() + (double)f1, this.getZ() + vector3d.z);
+            passenger.setYRot(passenger.getYRot() + this.deltaRotation);
             passenger.setYHeadRot(passenger.getYHeadRot() + this.deltaRotation);
-            applyYawToEntity(passenger);
+            applyOriantationsToEntity(passenger);
         }
     }
 }
