@@ -4,10 +4,10 @@ import com.talhanation.smallships.Main;
 import com.talhanation.smallships.entities.BriggEntity;
 import com.talhanation.smallships.entities.CogEntity;
 import com.talhanation.smallships.entities.projectile.CannonBallEntity;
-import net.minecraft.entity.EntityClassification;
-import net.minecraft.entity.EntityType;
-import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.fml.RegistryObject;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.MobCategory;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -16,20 +16,20 @@ public class ModEntityTypes {
     public static final DeferredRegister<EntityType<?>> ENTITY_TYPES = DeferredRegister.create(ForgeRegistries.ENTITIES, Main.MOD_ID);
 
     public static final RegistryObject<EntityType<CogEntity>> COG = ENTITY_TYPES.register("cog",
-            () -> EntityType.Builder.<CogEntity>of(CogEntity::new, EntityClassification.MISC)
+            () -> EntityType.Builder.<CogEntity>of(CogEntity::new, MobCategory.MISC)
                     .sized(3.5F, 1.25F)
                     .clientTrackingRange(20)
                     .setUpdateInterval(10)
                     .setShouldReceiveVelocityUpdates(true)
-                    .build(new ResourceLocation(Main.MOD_ID, "cog").toString()));
+                    .build(ResLocation.COG.toString()));
 
     public static final RegistryObject<EntityType<BriggEntity>> BRIGG = ENTITY_TYPES.register("brigg",
-            () -> EntityType.Builder.<BriggEntity>of(BriggEntity::new, EntityClassification.MISC)
+            () -> EntityType.Builder.<BriggEntity>of(BriggEntity::new, MobCategory.MISC)
                     .sized(3.5F, 1.25F)
                     .clientTrackingRange(20)
                     .setUpdateInterval(10)
                     .setShouldReceiveVelocityUpdates(true)
-                    .build(new ResourceLocation(Main.MOD_ID, "brigg").toString()));
+                    .build(ResLocation.BRIGG.toString()));
 
 /*
     public static final RegistryObject<EntityType<SailShipPart>> SHIP_PART = ENTITY_TYPES.register("ship_part",
@@ -41,10 +41,16 @@ public class ModEntityTypes {
                     .build(new ResourceLocation(Main.MOD_ID, "ship_part").toString()));
     */
     public static final RegistryObject<EntityType<CannonBallEntity>> CANNON_BALL = ENTITY_TYPES.register("cannon_ball",
-            () -> EntityType.Builder.<CannonBallEntity>of(CannonBallEntity::new, EntityClassification.MISC)
+            () -> EntityType.Builder.<CannonBallEntity>of(CannonBallEntity::new, MobCategory.MISC)
                     .sized(0.25F, 0.25F)
                     .clientTrackingRange(20)
                     .setUpdateInterval(10)
                     .setShouldReceiveVelocityUpdates(true)
-                    .build(new ResourceLocation(Main.MOD_ID, "cannon_ball").toString()));
+                    .build(ResLocation.CANNON_BALL.toString()));
+
+    public static class ResLocation {
+        public static final ResourceLocation COG = new ResourceLocation(Main.MOD_ID, "cog");
+        public static final ResourceLocation BRIGG = new ResourceLocation(Main.MOD_ID, "brigg");
+        public static final ResourceLocation CANNON_BALL = new ResourceLocation(Main.MOD_ID, "cannon_ball");
+    }
 }
