@@ -1,35 +1,52 @@
 package com.talhanation.smallships.items;
 
 import com.talhanation.smallships.client.render.RenderItemCog;
-import com.talhanation.smallships.entities.CogEntity;
 import com.talhanation.smallships.entities.AbstractWaterVehicle;
+import com.talhanation.smallships.entities.CogEntity;
 import com.talhanation.smallships.init.ModEntityTypes;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.stats.Stats;
-import net.minecraft.util.*;
-import net.minecraft.util.math.AxisAlignedBB;
-import net.minecraft.util.math.RayTraceContext;
-import net.minecraft.util.math.RayTraceResult;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.world.World;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntitySelector;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.ClipContext;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.AABB;
+import net.minecraft.world.phys.HitResult;
+import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.client.IItemRenderProperties;
 
 import java.util.List;
+import java.util.function.Consumer;
 import java.util.function.Predicate;
 
-
 public class CogItem extends Item {
-    private static final Predicate<Entity> X = EntityPredicates.NO_SPECTATORS.and(Entity::canBeCollidedWith);
+    private static final Predicate<Entity> X = EntitySelector.NO_SPECTATORS.and(Entity::canBeCollidedWith);
     private final CogEntity.Type type;
 
     public CogItem(CogEntity.Type typeIn, Item.Properties properties) {
-        super(properties.setISTER(() -> RenderItemCog::new));
+        super(properties);
         this.type = typeIn;
     }
 
-    public CogEntity getCogEntity(World world) {
+    @Override
+    public void initializeClient(Consumer<IItemRenderProperties> consumer) {
+        consumer.accept(new IItemRenderProperties() {
+            @Override
+            public BlockEntityWithoutLevelRenderer getItemStackRenderer() {
+                return new RenderItemCog(Minecraft.getInstance().getBlockEntityRenderDispatcher(), Minecraft.getInstance().getEntityModels());
+            }
+        });
+    }
+
+    public CogEntity getCogEntity(Level world) {
         CogEntity cog = new CogEntity(ModEntityTypes.COG.get(), world);
         cog.setWoodType(type);
         cog.setSailState(4);
@@ -42,49 +59,49 @@ public class CogItem extends Item {
 
 
     @Override
-    public ActionResult<ItemStack> use(World worldIn, PlayerEntity playerIn, Hand handIn) {
+    public InteractionResultHolder<ItemStack> use(Level worldIn, Player playerIn, InteractionHand handIn) {
         ItemStack itemstack = playerIn.getItemInHand(handIn);
-        RayTraceResult raytraceresult = getPlayerPOVHitResult(worldIn, playerIn, RayTraceContext.FluidMode.ANY);
-        if (raytraceresult.getType() == RayTraceResult.Type.MISS) {
-            return ActionResult.pass(itemstack);
+        HitResult raytraceresult = getPlayerPOVHitResult(worldIn, playerIn, ClipContext.Fluid.ANY);
+        if (raytraceresult.getType() == HitResult.Type.MISS) {
+            return InteractionResultHolder.pass(itemstack);
         } else {
-            Vector3d vector3d = playerIn.getViewVector(1.0F);
+            Vec3 vector3d = playerIn.getViewVector(1.0F);
             double d0 = 5.0D;
             List<Entity> list = worldIn.getEntities(playerIn, playerIn.getBoundingBox().expandTowards(vector3d.scale(5.0D)).inflate(1.0D), X);
             if (!list.isEmpty()) {
-                Vector3d vector3d1 = playerIn.getEyePosition(1.0F);
+                Vec3 vector3d1 = playerIn.getEyePosition(1.0F);
 
                 for(Entity entity : list) {
-                    AxisAlignedBB axisalignedbb = entity.getBoundingBox().inflate((double)entity.getPickRadius());
+                    AABB axisalignedbb = entity.getBoundingBox().inflate(entity.getPickRadius());
                     if (axisalignedbb.contains(vector3d1)) {
-                        return ActionResult.pass(itemstack);
+                        return InteractionResultHolder.pass(itemstack);
                     }
                 }
             }
 
-            if (raytraceresult.getType() == RayTraceResult.Type.BLOCK) {
+            if (raytraceresult.getType() == HitResult.Type.BLOCK) {
                 CogEntity boatentity = new CogEntity(worldIn, raytraceresult.getLocation().x, raytraceresult.getLocation().y, raytraceresult.getLocation().z);
                 boatentity.setWoodType(this.type);
-                boatentity.yRot = playerIn.yRot + 90F;
+                boatentity.setYRot(playerIn.getYRot() + 90F);
                 if (!worldIn.noCollision(boatentity, boatentity.getBoundingBox().inflate(-0.1D))) {
-                    return ActionResult.fail(itemstack);
+                    return InteractionResultHolder.fail(itemstack);
                 } else {
                     if (!worldIn.isClientSide) {
                         worldIn.addFreshEntity(boatentity);
                         if (boatentity.getStatus().equals(AbstractWaterVehicle.Status.IN_WATER)) {
-                            worldIn.playSound(null, boatentity.getX(), boatentity.getY(), boatentity.getZ(), SoundEvents.PLAYER_SPLASH, SoundCategory.BLOCKS, 0.75F, 0.8F);
-                            worldIn.playSound(null, boatentity.getX(), boatentity.getY(), boatentity.getZ(), SoundEvents.AMBIENT_UNDERWATER_ENTER, SoundCategory.BLOCKS, 0.75F, 0.8F);
+                            worldIn.playSound(null, boatentity.getX(), boatentity.getY(), boatentity.getZ(), SoundEvents.PLAYER_SPLASH, SoundSource.BLOCKS, 0.75F, 0.8F);
+                            worldIn.playSound(null, boatentity.getX(), boatentity.getY(), boatentity.getZ(), SoundEvents.AMBIENT_UNDERWATER_ENTER, SoundSource.BLOCKS, 0.75F, 0.8F);
                         }
-                        if (!playerIn.abilities.instabuild) {
+                        if (!playerIn.getAbilities().instabuild) {
                             itemstack.shrink(1);
                         }
                     }
 
                     playerIn.awardStat(Stats.ITEM_USED.get(this));
-                    return ActionResult.sidedSuccess(itemstack, worldIn.isClientSide());
+                    return InteractionResultHolder.sidedSuccess(itemstack, worldIn.isClientSide());
                 }
             } else {
-                return ActionResult.pass(itemstack);
+                return InteractionResultHolder.pass(itemstack);
             }
         }
     }
